Rename truffle-migrate to @truffle/migrate
