+++ packages/core/lib/commands/migrate.js
@@ -144,7 +144,7 @@ const command = {
   prepareConfigForRealMigrations: async function(buildDir, options) {
     const Artifactor = require("@truffle/artifactor");
     const Resolver = require("truffle-resolver");
-    const Migrate = require("truffle-migrate");
+    const Migrate = require("@truffle/migrate");
     const { Environment } = require("@truffle/environment");
     const Config = require("@truffle/config");
 
@@ -177,7 +177,7 @@ const command = {
   run: function(options, done) {
     const Artifactor = require("@truffle/artifactor");
     const Resolver = require("truffle-resolver");
-    const Migrate = require("truffle-migrate");
+    const Migrate = require("@truffle/migrate");
     const Contracts = require("@truffle/workflow-compile");
     const { Environment } = require("@truffle/environment");
     const Config = require("@truffle/config");
