+++ packages/core/package.json
@@ -60,7 +60,7 @@
     "@truffle/provider": "^0.1.16",
     "@truffle/provisioner": "^0.1.5",
     "@truffle/interface-adapter": "^0.2.5",
-    "truffle-migrate": "^3.0.32",
+    "@truffle/migrate": "^3.0.32",
     "truffle-resolver": "^5.0.16",
     "universal-analytics": "^0.4.17",
     "web3": "1.2.1",
