+++ packages/debugger/package.json
@@ -67,7 +67,7 @@
     "@truffle/debug-utils": "^1.0.18",
     "@truffle/workflow-compile": "^2.1.3",
     "@truffle/artifactor": "^4.0.30",
-    "truffle-migrate": "^3.0.32",
+    "@truffle/migrate": "^3.0.32",
     "truffle-resolver": "^5.0.16",
     "webpack": "^3.8.1",
     "webpack-dev-middleware": "^2.0.4",
