+++ packages/debugger/test/helpers.js
@@ -7,7 +7,7 @@ import Contracts from "@truffle/workflow-compile";
 import Debug from "@truffle/debug-utils";
 import Artifactor from "@truffle/artifactor";
 import Web3 from "web3";
-import Migrate from "truffle-migrate";
+import Migrate from "@truffle/migrate";
 import Box from "@truffle/box";
 import Resolver from "truffle-resolver";
