+++ packages/migrate/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "truffle-migrate",
+  "name": "@truffle/migrate",
   "description": "On-chain migrations management",
   "license": "MIT",
   "author": "Tim Coulter <tim@trufflesuite.com>",
-  "homepage": "https://github.com/trufflesuite/truffle-migrate#readme",
-  "repository": "https://github.com/trufflesuite/truffle/tree/master/packages/truffle-migrate",
+  "homepage": "https://github.com/trufflesuite/truffle/tree/master/packages/migrate#readme",
+  "repository": "https://github.com/trufflesuite/truffle/tree/master/packages/migrate",
   "bugs": {
-    "url": "https://github.com/trufflesuite/truffle-migrate/issues"
+    "url": "https://github.com/trufflesuite/truffle/issues"
   },
   "version": "3.0.32",
   "main": "index.js",
